wip migrate scripts
