+++ stellar-engine/src/stellar_engine/entities/inputs.py
